requirements.txt
streamlit==1.34.0
streamlit-webrtc==0.45.0
pydub
av
requests
python-dotenv